make single cloud detection function with two methods available (dec/vg)
