+++ aprofiles/detection/clouds.py
@@ -1,34 +1,35 @@
 # @author Augustin Mortier
 # @desc A-Profiles - Clouds detection
 
-import numpy as np
 from rich.progress import track
+from typing import Literal
 
+import aprofiles.profiles
 from aprofiles import utils
 
 
-def detect_clouds(profiles, time_avg=1., zmin=0., thr_noise=5., thr_clouds=4., min_snr=0., verbose=False):
+def detect_clouds(profiles: aprofiles.profiles.ProfilesData, method: Literal["dec", "vg"]="dec", time_avg: float=1., zmin: float=0., thr_noise: float=5., thr_clouds: float=4., min_snr: float=0., verbose: bool=False):
     """Module for *clouds detection*.
-    The detection is performed on each individual profile. It is based on the analysis of the vertical gradient of the profile as respect to the level of noise measured in the profile.
+    Two methods are available: 
+    - "dec" (default): Deep Embedded Clustering (see [AI-Profiles](https://github.com/AugustinMortier/ai-profiles)).
+    - "vg": Vertical Gradient.
 
     Args:
         profiles (aprofiles.profiles.ProfilesData): `ProfilesData` instance.
+        method (Literal["dec", "vg"]): cloud detection method used.
         time_avg (float, optional): in minutes, the time during which we aggregates the profiles prior to the clouds detection.
-        zmin (float, optional): altitude AGL, in m, above which we look for clouds. We recommend using the same value as used in the extrapolation_low_layers method.
-        thr_noise (float, optional): threshold used in the test to determine if a couple (base,peak) is significant: data[peak(z)] data[base(z)] >= thr_noise * noise(z).
-        thr_clouds (float, optional): threshold used to discriminate aerosol from clouds: data[peak(z)] / data[base(z)] >= thr_clouds.
-        min_snr (float, optional): Minimum SNR required at the clouds peak value to consider the cloud as valid.
+        zmin (float, optional): altitude AGL, in m, above which we look for clouds. We recommend using the same value as used in the extrapolation_low_layers method (only for 'vg' method).
+        thr_noise (float, optional): threshold used in the test to determine if a couple (base,peak) is significant: data[peak(z)] data[base(z)] >= thr_noise * noise(z) (only for 'vg' method).
+        thr_clouds (float, optional): threshold used to discriminate aerosol from clouds: data[peak(z)] / data[base(z)] >= thr_clouds (only for 'vg' method).
+        min_snr (float, optional): Minimum SNR required at the clouds peak value to consider the cloud as valid (only for 'vg' method).
         verbose (bool, optional): verbose mode. Defaults to `False`.
 
     Returns:
         (aprofiles.profiles.ProfilesData): 
             adds the following (xarray.DataArray) to existing (aprofiles.profiles.ProfilesData):
+            - `'clouds' (time, altitude)`: Mask array corresponding to data flagged as a cloud.
 
-            - `'clouds_bases' (time, altitude)`: Mask array corresponding to the bases of the clouds.
-            - `'clouds_peaks' (time, altitude)`: Mask array corresponding to the peaks (maximum signal measured) of the clouds.
-            - `'clouds_tops' (time, altitude)`: Mask array corresponding to the top of the cloud if the beam crosses the cloud. If not, the top corresponds to the first value where the signal becomes lower than the one measured at the base of the cloud.
-
-    Example:
+    Example 1:
         ```python
         import aprofiles as apro
         # read example file
@@ -36,7 +37,22 @@ def detect_clouds(profiles, time_avg=1., zmin=0., thr_noise=5., thr_clouds=4., m
         reader = apro.reader.ReadProfiles(path)
         profiles = reader.read()
         # clouds detection
-        profiles.clouds(zmin=300.)
+        profiles.clouds(method="dec")
+        # attenuated backscatter image with clouds
+        profiles.plot(show_ml_clouds=True, vmin=1e-2, vmax=1e1, log=True)
+        ```
+
+        ![Clouds detection](../../assets/images/ml_clouds.png)
+    
+    Example 2:
+        ```python
+        import aprofiles as apro
+        # read example file
+        path = "examples/data/L2_0-20000-001492_A20210909.nc"
+        reader = apro.reader.ReadProfiles(path)
+        profiles = reader.read()
+        # clouds detection
+        profiles.clouds(method="vg", zmin=300.)
         # attenuated backscatter image with clouds
         profiles.plot(show_clouds=True, vmin=1e-2, vmax=1e1, log=True)
         ```
@@ -44,7 +60,7 @@ def detect_clouds(profiles, time_avg=1., zmin=0., thr_noise=5., thr_clouds=4., m
         ![Clouds detection](../../assets/images/clouds.png)
     """
 
-    def _detect_clouds_from_rcs(data, zmin, thr_noise, thr_clouds, min_snr):
+    def _vg_clouds(data, zmin, thr_noise, thr_clouds, min_snr):
         # data: 1D range corrected signal (rcs)
         # zmin: altitude AGL, in m, above which we detect clouds
         # thr_noise: threshold used in the test to determine if a couple (base,peak) is significant: data[peak(z)] - data[base(z)] >= thr_noise * noise(z)
@@ -159,6 +175,24 @@ def detect_clouds(profiles, time_avg=1., zmin=0., thr_noise=5., thr_clouds=4., m
                     peaks[i_peaks[i]] = False
             # it is important to keep the tops in the same order, so not to use utils.get_true_indexes() function here
             return utils.get_true_indexes(bases), utils.get_true_indexes(peaks), i_tops
+
+        def _mask_cloud(profile: aprofiles.profiles, bases: list[int], tops: list[int]) -> np.ndarray:
+            """Generate a boolean mask for cloud presence in a given profile.
+
+            Args:
+                profile (aprofiles.profiles): 1D array representing a single atmospheric profile.
+                bases (list[int]): List of base indices where clouds start.
+                tops (list[int]): List of top indices where clouds end.
+
+            Returns:
+                np.ndarray: Boolean mask of the same length as profile, where cloud regions are True.
+            """
+            mask = np.zeros_like(profile, dtype=bool)
+            
+            for base, top in zip(bases, tops):
+                mask[base:top] = True
+            
+            return mask
 
         #-1. check if any valid data
         if np.isnan(data).all():
@@ -285,8 +319,8 @@ def detect_clouds(profiles, time_avg=1., zmin=0., thr_noise=5., thr_clouds=4., m
         i_peaks = [i for i in i_peaks if i not in remove_peaks]
         i_tops = [i for i in i_tops if i not in remove_tops]
 
-        # 11. rebuild masks from indexes
-        bases, peaks, tops = _make_all_masks(data, i_bases, i_peaks, i_tops)
+        # 11. build cloud mask from indexes
+        clouds = _mask_cloud(data, i_bases, i_tops)
 
         """
         #some plotting
@@ -320,59 +354,187 @@ def detect_clouds(profiles, time_avg=1., zmin=0., thr_noise=5., thr_clouds=4., m
         fig.suptitle(t,weight='bold')
         """
 
-        return {
-            "bases": bases,
-            "peaks": peaks,
-            "tops": tops,
+        return clouds
+
+    
+    def _prepare_data(data, vmin, vmax, target_size):
+        # Log transform
+        #data = np.log(data)
+
+        # Clip data to the range [vmin, vmax] and then scale it
+        data_clipped = np.clip(data, vmin, vmax)
+
+        # Replace NaNs with minimum values
+        np.nan_to_num(data_clipped, copy=False, nan=vmin)
+
+        # Invert the normalized data to reflect the 'gray_r' colormap
+        data_normalized = (data_clipped - vmin) / (vmax - vmin)
+        data_inverted = 1 - data_normalized  # Invert the grayscale values
+        
+        # Resize the inverted data to the target size
+        resized_data = resize(
+            data_inverted,
+            output_shape=target_size,
+            order=0,  # Nearest-neighbor interpolation to avoid smoothing
+            anti_aliasing=False
+        )
+        return resized_data
+    
+    
+    def _ml_clouds(prepared_data, encoder, cluster, target_size, output_shape):
+        # Add batch and channel dimensions for further processing
+        data_array = np.expand_dims(np.expand_dims(prepared_data, axis=0), axis=-1)
+
+        # Encode the image to get feature representation
+        encoded_img = encoder.predict(data_array)[0]  # Remove batch dimension
+
+        # Step 1: Aggregate Encoded Features
+        aggregated_encoded_img = np.mean(encoded_img, axis=-1)  # Aggregated to single-channel (16, 32)
+
+        # Optional: Normalize for better visualization
+        aggregated_encoded_img = (aggregated_encoded_img - aggregated_encoded_img.min()) / (aggregated_encoded_img.max() - aggregated_encoded_img.min())
+
+        # Step 2: Flatten Encoded Features and Cluster
+        encoded_img_flat = encoded_img.reshape(-1, encoded_img.shape[-1])  # Flatten spatial dimensions for clustering
+        pixel_labels = cluster.predict(encoded_img_flat)  # Get cluster labels for each pixel
+
+        # Step 3: Map clusters to categories
+        category_mapping = {
+            1: False,  # molecules
+            3: False,  # molecules
+            5: False,  # noise
+            4: False,  # aerosols
+            2: True,  # clouds
+            6: True,  # clouds
+            0: False,  # other
+            7: False   # other
         }
+        pixel_labels = np.vectorize(category_mapping.get)(pixel_labels)
+
+        # Reshape the cluster labels back to the spatial dimensions
+        pixel_labels_image_shape = pixel_labels.reshape(encoded_img.shape[0], encoded_img.shape[1])
+
+        # Step 4: Upsample the cluster labels to match the original image size
+        upsampled_pixel_labels = resize(
+            pixel_labels_image_shape,
+            (target_size[0], target_size[1]),
+            order=0,  # Nearest-neighbor interpolation
+            preserve_range=True,
+            anti_aliasing=False
+        )
+        #upsampled_pixel_labels = upsampled_pixel_labels.astype(int)  # Ensure the labels are integers
+        
+        # resize upsampled_pixel_labels to original size
+        resized_upsampled_pixel_labels = resize(
+                upsampled_pixel_labels,
+                output_shape=output_shape,
+                order=0,  # Nearest-neighbor interpolation to avoid smoothing
+                anti_aliasing=False
+            )
+        
+        return resized_upsampled_pixel_labels
+    
+    def _split_matrix(matrix, max_size):
+        return [matrix[i:i+max_size] for i in range(0, matrix.shape[0], max_size)]
+
+    def _combine_matrices(matrices):
+        return np.vstack(matrices)
+
 
     # we work on profiles averaged in time to reduce the noise
     rcs = profiles.time_avg(
         time_avg, var="attenuated_backscatter_0"
     ).data.attenuated_backscatter_0
 
-    clouds_bases, clouds_peaks, clouds_tops = [], [], []
-    for i in (track(range(len(profiles.data.time.data)), description="clouds",disable=not verbose)):
-        clouds = _detect_clouds_from_rcs(
-            rcs.data[i, :], zmin, thr_noise, thr_clouds, min_snr
-        )
+    # imports and check options
+    if method == 'dec':
+        import os
+        import warnings
+        import absl.logging
+        from sklearn.exceptions import InconsistentVersionWarning
 
-        # store info in 2D array
-        clouds_bases.append(clouds["bases"])
-        clouds_peaks.append(clouds["peaks"])
-        clouds_tops.append(clouds["tops"])
+        # Suppress TensorFlow CUDA messages
+        os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
+        os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
+        os.environ['XLA_FLAGS'] = '--xla_gpu_cuda_data_dir='  # Prevents unnecessary CUDA checks
+        os.environ["TF_FORCE_GPU_ALLOW_GROWTH"] = "false"  # Avoids some GPU-related logs
+
+        # Suppress Abseil warnings
+        absl.logging.set_verbosity(absl.logging.ERROR)
+
+        # Suppress scikit-learn version warnings
+        warnings.simplefilter("ignore", InconsistentVersionWarning)
+
+        # Suppress TensorFlow progress bar
+        import tensorflow as tf
+        tf.keras.utils.disable_interactive_logging()
+
+        import numpy as np
+        from tensorflow.keras.models import load_model
+        from skimage.transform import resize
+        from pathlib import Path
+        import joblib
+        
+        # ML parameters
+        ML = {
+            'paths': {
+                'encoder': Path(Path(__file__).parent,'ml_models','encoder.keras'),
+                'kmeans': Path(Path(__file__).parent,'ml_models','kmeans.pkl')    
+            },
+            'params': {
+                'vmin': -2,
+                'vmax': 2,
+                'target_size': (256, 512)    
+            }
+        }
+        
+        # we split the rcs into subsets, so we artificially increase the time resolution of the cloud detection which is degraded by the DEC technique.
+        split_rcs_list = _split_matrix(rcs, 100)
+        
+        # Load the encoder and kmeans model
+        encoder = load_model(ML['paths']['encoder'])
+        cluster = joblib.load(ML['paths']['kmeans'])
+        
+        # prepare data
+        split_ml_clouds = []
+        for split_rcs in split_rcs_list:
+            prepared_data = _prepare_data(split_rcs, ML['params']['vmin'], ML['params']['vmax'], ML['params']['target_size'])
+            split_ml_clouds.append(_ml_clouds(prepared_data, encoder, cluster, ML['params']['target_size'], output_shape=np.shape(split_rcs)))
+        
+        # aggregate split_ml_clouds
+        clouds = _combine_matrices(split_ml_clouds)
+        
+    elif method == 'vg':
+        import numpy as np
+        if None in (zmin, thr_noise, thr_clouds, min_snr):
+            raise ValueError("For method 'vg', zmin, thr_noise, thr_clouds, and min_snr must be provided.")
+
+        clouds = []
+        for i in (track(range(len(profiles.data.time.data)), description="clouds",disable=not verbose)):
+            i_clouds = _vg_clouds(
+                rcs.data[i, :], zmin, thr_noise, thr_clouds, min_snr
+            )
+
+            # store info in 2D array
+            clouds.append(i_clouds)
 
     # creates dataarrays
-    profiles.data["clouds_bases"] = (("time", "altitude"), clouds_bases)
-    profiles.data["clouds_bases"] = profiles.data.clouds_bases.assign_attrs({
-        'long_name': 'Mask - Base height of clouds',
+    profiles.data["clouds"] = (("time", "altitude"), clouds)
+    profiles.data["clouds"] = profiles.data.clouds.assign_attrs({
+        'long_name': 'Clouds mask',
         'units': 'bool',
         'time_avg': time_avg,
-        'thr_noise': thr_noise,
-        'thr_clouds': thr_clouds,
-    })
-
-    profiles.data["clouds_peaks"] = (("time", "altitude"), clouds_peaks)
-    profiles.data["clouds_peaks"] = profiles.data.clouds_peaks.assign_attrs({
-        'long_name': 'Mask - Peak height of clouds',
-        'units': 'bool',
-        'time_avg': time_avg,
-        'thr_noise': thr_noise,
-        'thr_clouds': thr_clouds,
-    })
-
-    profiles.data["clouds_tops"] = (("time", "altitude"), clouds_tops)
-    profiles.data["clouds_tops"] = profiles.data.clouds_tops.assign_attrs({
-        'long_name': 'Mask - Top height of clouds',
-        'units': 'bool',
-        'time_avg': time_avg,
-        'thr_noise': thr_noise,
-        'thr_clouds': thr_clouds,
+        'method': method,
+        'options': {
+            'thr_noise': thr_noise,
+            'thr_clouds': thr_clouds,
+        }
     })
     return profiles
 
 
 def _main():
+    import numpy as np
     import aprofiles as apro
 
     path = "examples/data/E-PROFILE/L2_0-20000-001492_A20210909.nc"
@@ -381,13 +543,13 @@ def _main():
 
     # basic corrections
     profiles.extrapolate_below(z=150.0, inplace=True)
-    # profiles.desaturate_below(z=4000., inplace=True)
+    profiles.desaturate_below(z=4000., inplace=True)
 
     # detection
-    profiles.clouds(zmin=300, thr_noise=4, thr_clouds=4, verbose=True)
+    profiles.clouds(method="dec", zmin=300, thr_noise=4, thr_clouds=4, verbose=True)
     profiles.plot(show_clouds=True, log=True, vmin=1e-2, vmax=1e1)
     # plot single profile
-    datetime = np.datetime64("2021-09-09T14:00:00")
+    datetime = np.datetime64("2021-09-09T21:20:00")
     profiles.plot(datetime=datetime, vmin=-1, vmax=10, zmax=12000, show_clouds=True)
