+++ aprofiles/profiles.py
@@ -7,6 +7,7 @@ from pathlib import Path
 import numpy as np
 import xarray as xr
 from rich.progress import track
+from typing import Literal
 
 import aprofiles as apro
 
@@ -62,24 +63,13 @@ class ProfilesData:
             return (
                 min(np.diff(self.data.time.data)).astype("timedelta64[s]").astype(int)
             )
-
+    
     def _get_lowest_clouds(self):
         # returns an array of the altitude (in m, ASL) of the lowest cloud for each timestamp
-        lowest_clouds = []
-        for i in np.arange(len(self.data.time.data)):
-            i_clouds = apro.utils.get_true_indexes(self.data.clouds_bases.data[i, :])
-            if len(i_clouds) > 0:
-                lowest_clouds.append(self.data.altitude.data[i_clouds[0]])
-            else:
-                lowest_clouds.append(np.nan)
-        return lowest_clouds
-
-    def _get_lowest_ml_clouds(self):
-        # returns an array of the altitude (in m, ASL) of the lowest ml_cloud for each timestamp
         lowest_clouds = np.full(np.shape(self.data.time.data), np.nan)
         for i, _ in enumerate(self.data.time.data):
-            ml_clouds_profile = self.data.ml_clouds[i,:]
-            i_clouds = np.squeeze(np.where(ml_clouds_profile))
+            clouds_profile = self.data.clouds[i,:]
+            i_clouds = np.squeeze(np.where(clouds_profile))
             if len(i_clouds) > 0:
                 lowest_clouds[i] = self.data.altitude.data[i_clouds[0]]
         return lowest_clouds
@@ -398,18 +388,18 @@ class ProfilesData:
         Calls :meth:`aprofiles.detection.foc.detect_foc()`.
         """
         return apro.detection.foc.detect_foc(self, method, var, z_snr, min_snr, zmin_cloud)
-
-    def clouds(self, time_avg=1, zmin=0, thr_noise=5., thr_clouds=4., min_snr=0.0, verbose=False):
+    
+    def clouds(self, method: Literal["dec", "vg"]="dec", time_avg=1, zmin=0, thr_noise=5., thr_clouds=4., min_snr=0.0, verbose=False):
         """
         Calls :meth:`aprofiles.detection.clouds.detect_clouds()`.
         """
-        return apro.detection.clouds.detect_clouds(self, time_avg, zmin, thr_noise, thr_clouds, min_snr, verbose)
+        return apro.detection.clouds.detect_clouds(self, method, time_avg, zmin, thr_noise, thr_clouds, min_snr, verbose)
 
-    def ml_clouds(self, time_avg=1, verbose=False):
-        """
-        Calls :meth:`aprofiles.detection.ml_clouds.detect_clouds()`.
-        """
-        return apro.detection.ml_clouds.detect_clouds(self, time_avg, verbose)
+    #def ml_clouds(self, time_avg=1, verbose=False):
+    #    """
+    #    Calls :meth:`aprofiles.detection.ml_clouds.detect_clouds()`.
+    #    """
+    #    return apro.detection.ml_clouds.detect_clouds(self, time_avg, verbose)
     
     def pbl(self, time_avg=1, zmin=100., zmax=3000., wav_width=200., under_clouds=True, min_snr=2., verbose=False):
         """
@@ -431,7 +421,7 @@ class ProfilesData:
 
     def plot(
         self, var="attenuated_backscatter_0", datetime=None, zref="agl", zmin=None, zmax=None, vmin=None, vmax=None, log=False,
-        show_foc=False, show_pbl=False, show_clouds=False, show_ml_clouds=False, cmap="coolwarm", show_fig=True, save_fig=None, **kwargs
+        show_foc=False, show_pbl=False, show_clouds=False, cmap="coolwarm", show_fig=True, save_fig=None, **kwargs
     ):
         """
         Plotting method.
@@ -450,7 +440,6 @@ class ProfilesData:
             show_foc (bool, optional): Show fog or condensation retrievals.
             show_pbl (bool, optional): Show PBL height retrievals.
             show_clouds (bool, optional): Show clouds retrievals.
-            show_ml_clouds (bool, optional): Show ml_clouds retrievals.
             cmap (str, optional): Matplotlib colormap.
             show_fig (bool, optional): Show Figure.
             save_fig (str, optional): Path of the saved figure.
@@ -468,11 +457,11 @@ class ProfilesData:
         if datetime is None:
             # check dimension of var
             if len(list(self.data[var].dims)) == 2:
-                apro.plot.image.plot(self.data, var, zref, zmin, zmax, vmin, vmax, log, show_foc, show_pbl, show_clouds, show_ml_clouds, cmap, show_fig, save_fig)
+                apro.plot.image.plot(self.data, var, zref, zmin, zmax, vmin, vmax, log, show_foc, show_pbl, show_clouds, cmap, show_fig, save_fig)
             else:
                 apro.plot.timeseries.plot(self.data, var, show_fig, save_fig, **kwargs)
         else:
-            apro.plot.profile.plot(self.data, datetime, var, zref, zmin, zmax, vmin, vmax, log, show_foc, show_pbl, show_clouds, show_ml_clouds, show_fig, save_fig)
+            apro.plot.profile.plot(self.data, datetime, var, zref, zmin, zmax, vmin, vmax, log, show_foc, show_pbl, show_clouds, show_fig, save_fig)
     
     def write(self, base_dir=Path('examples', 'data', 'V-Profiles'), verbose=False):
         """
